requirements.txt
loguru~=0.7.0

pytube~=15.0.0

python-telegram-bot==13.7


